Update deltatable.py_Error Handling added to _apply_partial_logs function to handle schemaString error due to imperfection in deltalogs
While I am trying to load my delta table, I am facing key error. This error occured due to imperfections in delta logs as it can not find schemaString in one of the checkpoint in my delta logs. Due to this error DeltaTable is not able to load records. Using this error handling I am able to skip the problematic record and able to load the new records.